requirements/base.txt
# Base requirements

# Dash and friends.
dash
dash-renderer
dash-html-components
dash-core-components
dash-bootstrap-components
dash_table
plotly

# Numpy, Pandas and friends.
numpy
pandas
anndata
scanpy

# Caching functionality for Flask.
flask-caching

# Lightweight setup of data classes.
attrs
# Simplified logging.
logzero
# Support for YAML access.
ruamel.yaml
# Unified access to TAR and ZIP files.
fs
